@@ -1,7 +1,6 @@
 import numpy as np
 import dask
 import dask.array as da
-from daskms.optimisation import inlined_array
 from ducc0.wgridder.experimental import vis2dirty, dirty2vis
 from ducc0.misc import make_noncritical
 from uuid import uuid4
@@ -47,8 +46,6 @@ def hessian_xds(x, xds, hessopts, wsum, sigmainv, mask,
 
         convim = hessian(x[b], uvw, wgt, vis_mask, freq, beam, hessopts)
 
-        # convim = inlined_array(convim, uvw)
-
         convims[b] += convim
 
     convim = da.stack(convims)/wsum
@@ -66,7 +63,7 @@ def _hessian_impl(x, uvw, weight, vis_mask, freq, beam,
                   x0=0.0,
                   y0=0.0,
                   cell=None,
-                  wstack=None,
+                  do_wgridding=None,
                   epsilon=None,
                   double_accum=None,
                   nthreads=None):
@@ -83,7 +80,7 @@ def _hessian_impl(x, uvw, weight, vis_mask, freq, beam,
                     center_y=y0,
                     epsilon=epsilon,
                     nthreads=nthreads,
-                    do_wgridding=wstack,
+                    do_wgridding=do_wgridding,
                     divide_by_n=False)
 
     convim = vis2dirty(uvw=uvw,
@@ -99,7 +96,7 @@ def _hessian_impl(x, uvw, weight, vis_mask, freq, beam,
                       center_y=y0,
                       epsilon=epsilon,
                       nthreads=nthreads,
-                      do_wgridding=wstack,
+                      do_wgridding=do_wgridding,
                       double_precision_accumulation=double_accum,
                       divide_by_n=False)
 
@@ -162,11 +159,11 @@ def _hessian_psf_slice(
 
 from pfb.operators.hessian import _hessian_impl
 class hessian_psf_slice(object):
-    def __init__(self, ds, nbasis, nmax, nthreads, sigmainv, cell, wstack, epsilon, double_accum):
+    def __init__(self, ds, nbasis, nmax, nthreads, sigmainv, cell, do_wgridding, epsilon, double_accum):
         self.nthreads = nthreads
         self.sigmainv = sigmainv
         self.cell = cell
-        self.wstack = wstack
+        self.do_wgridding = do_wgridding
         self.epsilon = epsilon
         self.double_accum = double_accum
         self.lastsize = ds.PSF.shape[-1]
@@ -244,7 +241,7 @@ class hessian_psf_slice(object):
                                         self.freq,
                                         None,
                                         cell=self.cell,
-                                        wstack=self.wstack,
+                                        do_wgridding=self.do_wgridding,
                                         epsilon=self.epsilon,
                                         double_accum=self.double_accum,
                                         nthreads=self.nthreads)
@@ -272,7 +269,7 @@ def hessian_psf_cube(
         psf_convolve_cube(xpad, xhat, xout, psfhat, lastsize, x*beam,
                           nthreads=nthreads)
     else:
-        xout = psf_convolve_cube(xpad, xhat, xout, psfhat, lastsize, x,
+        psf_convolve_cube(xpad, xhat, xout, psfhat, lastsize, x,
                           nthreads=nthreads)
 
     if beam is not None:
@@ -284,34 +281,12 @@ def hessian_psf_cube(
     return xout + x * sigmainv
 
 
-# from ducc0.fft import r2c, c2r
-# def hess_psf(xpad,    # preallocated array to store padded image
-#              xhat,    # preallocated array to store FTd image
-#              xout,    # preallocated array to store output image
-#              psfhat,
-#              lastsize,
-#              x,       # input image, not overwritten
-#              nthreads=1,
-#              sigmainv=1.0):
-#     _, _, nx, ny = x.shape
-#     xpad[...] = 0.0
-#     xpad[:, :, 0:nx, 0:ny] = x
-#     r2c(xpad, axes=(-2, -1), nthreads=nthreads,
-#         forward=True, inorm=0, out=xhat)
-#     xhat *= psfhat
-#     c2r(xhat, axes=(-2, -1), forward=False, out=xpad,
-#         lastsize=lastsize, inorm=2, nthreads=nthreads,
-#         allow_overwriting_input=True)
-#     xout[...] = xpad[:, :, 0:nx, 0:ny]
-#     return xout + sigmainv*x
-
-
 def hess_vis(xds,
              dds,
              xout,
              x,
              sigmainv=1.0,
-             wstack=True,
+             do_wgridding=True,
              nthreads=1,
              epsilon=1e-7,
              divide_by_n=False):
@@ -340,7 +315,7 @@ def hess_vis(xds,
                               center_x=x0,
                               center_y=y0,
                               epsilon=epsilon,
-                              do_wgridding=wstack,
+                              do_wgridding=do_wgridding,
                               nthreads=nthreads,
                               divide_by_n=divide_by_n)
 
@@ -362,7 +337,7 @@ def hess_vis(xds,
                                                 center_x=x0,
                                                 center_y=y0,
                                                 epsilon=epsilon,
-                                                do_wgridding=wstack,
+                                                do_wgridding=do_wgridding,
                                                 nthreads=nthreads,
                                                 divide_by_n=divide_by_n)
             xout[field][f't{t}b{b}'] += sigmainv * x[field][f't{t}b{b}']
