@@ -4,6 +4,7 @@ import dask.array as da
 from distributed import wait, get_client, as_completed
 from operator import getitem
 from ducc0.misc import make_noncritical
+from pfb.utils.misc import norm_diff
 import pyscilog
 log = pyscilog.get_logger('PD')
 
@@ -23,7 +24,7 @@ def primal_dual(
         tol=1e-5,
         maxit=1000,
         minit=10,
-        positivity=True,
+        positivity=1,
         report_freq=10,
         gamma=1.0,
         verbosity=1):
@@ -53,8 +54,11 @@ def primal_dual(
 
         # primal update
         x = xp - tau * (psi(2 * v - vp) + grad(xp))
-        if positivity:
+        if positivity == 1:
             x[x < 0.0] = 0.0
+        elif positivity == 2:
+            msk = np.any(x<=0, axis=0)
+            x[:, msk] = 0.0
 
         # convergence check
         eps = np.linalg.norm(x - xp) / np.linalg.norm(x)
@@ -83,6 +87,7 @@ def primal_dual(
     return x, v
 
 
+from pfb.prox.prox_21m import dual_update_numba
 def primal_dual_optimised(
         x,  # initial guess for primal variable
         v,  # initial guess for dual variable
@@ -91,28 +96,29 @@ def primal_dual_optimised(
         psiH,  # adjoint of psi
         L,  # spectral norm of Hessian
         prox,  # prox of regulariser
+        l1weight,
+        reweighter,
         grad,  # gradient of smooth term
         nu=1.0,  # spectral norm of psi
         sigma=None,  # step size of dual update
         mask=None,  # regions where mask is False will be masked
         tol=1e-5,
         maxit=1000,
-        minit=10,
-        positivity=True,
+        positivity=1,
         report_freq=10,
         gamma=1.0,
-        verbosity=1):
+        verbosity=1,
+        maxreweight=50):
+    # TODO - we can't use make_noncritical because it sometimes returns an
+    # array that is not explicitly c-contiguous. How does this impact
+    # performance?
     # initialise
     xp = x.copy()
-    xp = make_noncritical(xp)
+    # xp = make_noncritical(xp)
     vp = v.copy()
-    vp = make_noncritical(vp)
-    vtilde = np.zeros_like(v)
-    vtilde = make_noncritical(vtilde)
-    vout = np.zeros_like(v)
-    vout = make_noncritical(vout)
     xout = np.zeros_like(x)
-    xout = make_noncritical(xout)
+    # xout = make_noncritical(xout)
+
 
     # this seems to give a good trade-off between
     # primal and dual problems
@@ -124,46 +130,47 @@ def primal_dual_optimised(
 
     # start iterations
     eps = 1.0
-    k = 0
-    while (eps > tol or k < minit) and k < maxit:
-        # tmp prox variable
-        # vtilde = v + sigma * psiH(xp)
-        psiH(xp, vtilde)
-        ne.evaluate('v + sigma * vtilde', out=vtilde)  #, casting='same_kind')
-
-        # dual update
-        # v = vtilde - sigma * prox(vtilde / sigma, lam / sigma)
-        prox(vtilde, vout, lam, sigma)
-        ne.evaluate('vtilde - sigma*vout', out=v)  #, casting='same_kind')
-
-        # primal update
-        # x = xp - tau * (psi(2 * v - vp) + grad(xp))
-        ne.evaluate('2.0 * v - vp', out=vout)  #, casting='same_kind')
-        psi(vout, xout)
+    numreweight = 0
+    for k in range(maxit):
+        psiH(xp, v)
+        dual_update_numba(vp, v, lam, sigma=sigma, weight=l1weight)
+        ne.evaluate('2.0 * v - vp', out=vp)  #, casting='same_kind')
+        psi(vp, xout)
         xout += grad(xp)
         ne.evaluate('xp - tau * xout', out=x)  #, casting='same_kind')
 
-        if positivity:
+        if positivity == 1:
             x[x < 0.0] = 0.0
+        elif positivity == 2:
+            msk = np.any(x<=0, axis=0)
+            x[:, msk] = 0.0
 
         # convergence check
-        eps = np.linalg.norm(x - xp) / np.linalg.norm(x)
+        if x.any():
+            eps = norm_diff(x, xp)
+        else:
+            import pdb; pdb.set_trace()
+            eps = 1.0
+        if eps < tol:
+            if reweighter is not None and numreweight < maxreweight:
+                l1weight = reweighter(x)
+                numreweight += 1
+            else:
+                if numreweight >= maxreweight:
+                    print("Maximum reweighting steps reached", file=log)
+                break
 
         # copy contents to avoid allocating new memory
-        # this is not faster but it allows us to see where the time is spent
-        np.copyto(xp, x)
-        np.copyto(vp, v)
-        # xp[...] = x[...]
-        # vp[...] = v[...]
+        xp[...] = x[...]
+        vp[...] = v[...]
 
         if np.isnan(eps) or np.isinf(eps):
             import pdb; pdb.set_trace()
 
         if not k % report_freq and verbosity > 1:
             print(f"At iteration {k} eps = {eps:.3e}", file=log)
-        k += 1
 
-    if k == maxit:
+    if k == maxit-1:
         if verbosity:
             print(f"Max iters reached. eps = {eps:.3e}", file=log)
     else:
