@@ -1,4 +1,5 @@
 # flake8: noqa
+from pathlib import Path
 from contextlib import ExitStack
 from pfb.workers.main import cli
 import click
@@ -27,7 +28,10 @@ def model2comps(**kw):
     opts = OmegaConf.create(defaults)
     import time
     timestamp = time.strftime("%Y%m%d-%H%M%S")
-    pyscilog.log_to_file(f'model2comps_{timestamp}.log')
+    ldir = Path(opts.log_directory).resolve()
+    ldir.mkdir(parents=True, exist_ok=True)
+    pyscilog.log_to_file(f'{ldir}/model2comps_{timestamp}.log')
+    print(f'Logs will be written to {str(ldir)}/model2comps_{timestamp}.log', file=log)
 
     OmegaConf.set_struct(opts, True)
 
@@ -60,16 +64,13 @@ def _model2comps(**kw):
     import dask.array as da
     from africanus.constants import c as lightspeed
     from pfb.utils.fits import load_fits, data_from_header
-    from pfb.utils.misc import restore_corrs, model_from_comps
     from astropy.io import fits
-    from pfb.utils.misc import compute_context
+    from pfb.utils.misc import compute_context, fit_image_cube
     import xarray as xr
-    from sympy.utilities.lambdify import lambdify
-    from sympy.parsing.sympy_parser import parse_expr
 
     basename = f'{opts.output_filename}_{opts.product.upper()}'
-    dds_name = f'{basename}_{opts.postfix}.dds.zarr'
-    coeff_name = f'{basename}_{opts.postfix}.coeffs.zarr'
+    dds_name = f'{basename}_{opts.postfix}.dds'
+    coeff_name = f'{basename}_{opts.postfix}_{opts.model_name.lower()}.mds'
 
     if os.path.isdir(coeff_name):
         if opts.overwrite:
@@ -110,122 +111,35 @@ def _model2comps(**kw):
     # wsums = [da.zeros(1) for _ in range(opts.nband)]
     model = np.zeros((ntime, nband, nx, ny), dtype=np.float64)
     wsums = np.zeros((ntime, nband), dtype=np.float64)
-    mask = np.zeros((nx, ny), dtype=bool)
     for ds in dds:
         b = ds.bandid
         t = ds.timeid
         model[t, b] = getattr(ds, opts.model_name).values
         wsums[t, b] += ds.WSUM.values[0]
 
-    # model = da.stack(model)
-    # wsums = da.stack(wsums).squeeze()
-    # model, wsums = dask.compute(model, wsums)
+    if not opts.use_wsum:
+        wsums[...] = 1.0
+
+    if opts.min_val is not None:
+        model = np.where(model >= opts.min_val, model, 0.0)
+
     if not np.any(model):
-        raise ValueError('Model is empty')
+        raise ValueError(f'Model is empty or has no components above {opts.min_val}')
     radec = (dds[0].ra, dds[0].dec)
 
-    ref_freq = mfreqs[0]
-    ref_time = mtimes[0]
-
-    # interpolate model in frequency
-    mask = np.any(model, axis=(0,1))
-    Ix, Iy = np.where(mask)
-    ncomps = Ix.size
-
-    # components excluding zeros
-    beta = model[:, :, Ix, Iy]
-    if opts.spectral_poly_order is not None and nband > 1:
-        orderf = opts.spectral_poly_order
-        if orderf > nband:
-            raise ValueError("spectral-poly-order can't be larger than nband")
-        print(f"Fitting freq axis with polynomial of order {orderf}", file=log)
-        # we are given frequencies at bin centers, convert to bin edges
-        delta_freq = mfreqs[1] - mfreqs[0]
-        wlow = (mfreqs - delta_freq/2.0)/ref_freq
-        whigh = (mfreqs + delta_freq/2.0)/ref_freq
-        wdiff = whigh - wlow
-
-        # set design matrix for each component
-        Xfit = np.zeros([mfreqs.size, orderf])
-        for i in range(1, orderf+1):
-            Xfit[:, i-1] = (whigh**i - wlow**i)/(i*wdiff)
-
-        comps = np.zeros((ntime, orderf, Ix.size))
-        freq_fitted = True
-        for t in range(ntime):
-            dirty_comps = Xfit.T.dot(wsums[t, :, None]*beta[t])
-
-            hess_comps = Xfit.T.dot(wsums[t, :, None]*Xfit)
-
-            comps[t] = np.linalg.solve(hess_comps, dirty_comps)
-
-    else:
-        raise NotImplementedError("Interpolation is currently mandatory")
-        print("Not fitting frequency axis", file=log)
-        comps = beta
-        freq_fitted = False
-        orderf = mfreqs.size
-
-    if opts.temporal_poly_order is not None and ntime > 1:
-        ordert = opts.temporal_poly_order
-        if order > ntime:
-            raise ValueError("temporal-poly-order can't be larger than ntime")
-        print(f"Fitting time axis with polynomial of order {orderf}", file=log)
-        # we are given times at bin centers, convert to bin edges
-        delta_time = mtimes[1] - mtimes[0]
-        wlow = (mtimes - delta_times/2.0)/ref_time
-        whigh = (mfreqs + delta_freq/2.0)/ref_time
-        wdiff = whigh - wlow
-
-        # set design matrix for each component
-        Xfit = np.zeros([mtimes.size, ordert])
-        for i in range(1, ordert+1):
-            Xfit[:, i-1] = (whigh**i - wlow**i)/(i*wdiff)
-
-        if freq_fitted:
-            compsnu = comps.copy()
-        comps = np.zeros((ordert, opts.spectral_poly_order, Ix.size))
-        time_fitted = True
-        for b in range(orderf):
-            dirty_comps = Xfit.T.dot(wsums[:, b, None]*compsnu[b])
-
-            hess_comps = Xfit.T.dot(wsums[:, b, None]*Xfit)
-
-            comps[:, b] = np.linalg.solve(hess_comps, dirty_comps)
-    else:
-        if ntime > 1:
-            raise NotImplementedError("Time interpolation is currently mandatory")
-        print("Not fitting time axis", file=log)
-        comps = comps
-        time_fitted = False
-        ordert = mtimes.size
-
-
-    # construct symbolic expression
-    from sympy.abc import t, f
-    from sympy import symbols
-
-    thetasf = []
-    params = ()
-    for i in range(orderf):
-        coefft = symbols(f't(0:{ordert})_f{i}')
-        # the reshape on comps needs to be consistent with the ordering in params
-        params += coefft
-        thetaf = sum(co*t**j for j, co in enumerate(coefft))
-        thetasf.append(thetaf)
-
-    polysym = sum(co*f**j for j, co in enumerate(thetasf))
-
+    coeffs, Ix, Iy, expr, params, texpr, fexpr = \
+        fit_image_cube(mtimes, mfreqs, model, wgt=wsums,
+                       nbasisf=opts.nbasisf, method=opts.fit_mode)
 
     # save interpolated dataset
     data_vars = {
-        'coefficients': (('params', 'comps'), comps.reshape(ordert*orderf, ncomps))
+        'coefficients': (('params', 'comps'), coeffs),
     }
     coords = {
         'location_x': (('location_x',), Ix),
         'location_y': (('location_y',), Iy),
         # 'shape_x':,
-        'params': (('params',), list(map(str, params))),
+        'params': (('params',), params),  # already converted to list
         'times': (('times',), mtimes),  # to allow rendering to original grid
         'freqs': (('freqs',), mfreqs)
     }
@@ -235,14 +149,14 @@ def _model2comps(**kw):
         'cell_rad_y': cell_rad,
         'npix_x': nx,
         'npix_y': ny,
-        'ref_freq': ref_freq,
-        'ref_time': ref_time,
+        'texpr': texpr,
+        'fexpr': fexpr,
         'center_x': x0,
         'center_y': y0,
         'ra': dds[0].ra,
         'dec': dds[0].dec,
         'stokes': opts.product,  # I,Q,U,V, IQ/IV, IQUV
-        'parametrisation': str(polysym)
+        'parametrisation': expr  # already converted to str
     }
 
 
@@ -250,9 +164,9 @@ def _model2comps(**kw):
                                coords=coords,
                                attrs=attrs)
     writes = xds_to_zarr(coeff_dataset,
-                         f'{basename}_{opts.postfix}.coeffs.zarr',
+                         coeff_name,
                          columns='ALL')
-    print(f'Writing interpolated model to {basename}_{opts.postfix}.coeffs.zarr',
+    print(f'Writing interpolated model to {coeff_name}',
           file=log)
     dask.compute(writes)
 
