@@ -53,7 +53,7 @@ def _restore(**kw):
     from pfb.utils.misc import Gaussian2D, fitcleanbeam, convolve2gaussres, dds2cubes
 
     basename = f'{opts.output_filename}_{opts.product.upper()}'
-    dds_name = f'{basename}_{opts.postfix}.dds.zarr'
+    dds_name = f'{basename}_{opts.postfix}.dds'
 
     dds = xds_from_zarr(dds_name)
     nband = opts.nband
@@ -68,8 +68,12 @@ def _restore(**kw):
         assert ds.y.size == ny
     freq = np.unique(np.array(freq))
     assert freq.size == opts.nband
-    nx_psf = dds[0].x_psf.size
-    ny_psf = dds[0].y_psf.size
+
+    # init fits headers
+    radec = (dds[0].ra, dds[0].dec)
+    ref_freq = np.mean(freq)
+    hdr_mfs = set_wcs(cell_deg, cell_deg, nx, ny, radec, ref_freq)
+    hdr = set_wcs(cell_deg, cell_deg, nx, ny, radec, freq)
 
 
     dirty, model, residual, psf, _, _, wsums, _ = dds2cubes(dds,
@@ -77,96 +81,126 @@ def _restore(**kw):
                                                             apparent=True)
     wsum = np.sum(wsums)
     output_type = dirty.dtype
-    psf_mfs = np.sum(psf, axis=0)
-    residual_mfs = np.sum(residual, axis=0)
     fmask = wsums > 0
+
+    if residual is None:
+        print('Warning, no residual in dds. '
+              'Using dirty as residual.', file=log)
+        residual = dirty.copy()
+    residual_mfs = np.sum(residual, axis=0)
     residual[fmask] /= wsums[fmask, None, None]/wsum
-    psf[fmask] /= wsums[fmask, None, None]/wsum
-    # sanity check
-    assert (psf_mfs.max() - 1.0) < 2e-7
-    assert ((np.amax(psf, axis=(1,2)) - 1.0) < 2e-7).all()
-
-    # fit restoring psf
-    GaussPar = fitcleanbeam(psf_mfs[None], level=0.5, pixsize=1.0)
-    GaussPars = fitcleanbeam(psf, level=0.5, pixsize=1.0)  # pixel units
-
-    cpsf_mfs = np.zeros(residual_mfs.shape, dtype=output_type)
-    cpsf = np.zeros(residual.shape, dtype=output_type)
-
-    lpsf = -(nx//2) + np.arange(nx)
-    mpsf = -(ny//2) + np.arange(ny)
-    xx, yy = np.meshgrid(lpsf, mpsf, indexing='ij')
-
-    cpsf_mfs = Gaussian2D(xx, yy, GaussPar[0], normalise=False)
-
-    for v in range(opts.nband):
-        cpsf[v] = Gaussian2D(xx, yy, GaussPars[v], normalise=False)
 
     if not model.any():
         print("Warning - model is empty", file=log)
     model_mfs = np.mean(model, axis=0)
 
-    image_mfs = convolve2gaussres(model_mfs[None], xx, yy,
-                                  GaussPar[0], opts.nvthreads,
-                                  norm_kernel=False)[0]  # peak of kernel set to unity
-    image_mfs += residual_mfs
-    image = np.zeros_like(model)
-    for b in range(nband):
-        image[b:b+1] = convolve2gaussres(model[b:b+1], xx, yy,
-                                         GaussPars[b], opts.nvthreads,
-                                         norm_kernel=False)  # peak of kernel set to unity
-        image[b] += residual[b]
+    if psf is not None:
+        nx_psf = dds[0].x_psf.size
+        ny_psf = dds[0].y_psf.size
+        psf_mfs = np.sum(psf, axis=0)
+        psf[fmask] /= wsums[fmask, None, None]/wsum
+        # sanity check
+        assert (psf_mfs.max() - 1.0) < 2e-7
+        assert ((np.amax(psf, axis=(1,2)) - 1.0) < 2e-7).all()
 
-    # convert pixel units to deg
-    GaussPar = list(GaussPar[0])
-    GaussPar[0] *= cell_deg
-    GaussPar[1] *= cell_deg
-    GaussPar = tuple(GaussPar)
+        # fit restoring psf
+        GaussPar = fitcleanbeam(psf_mfs[None], level=0.5, pixsize=1.0)
+        cpsf_mfs = np.zeros(residual_mfs.shape, dtype=output_type)
+        lpsf = -(nx//2) + np.arange(nx)
+        mpsf = -(ny//2) + np.arange(ny)
+        xx, yy = np.meshgrid(lpsf, mpsf, indexing='ij')
+        cpsf_mfs = Gaussian2D(xx, yy, GaussPar[0], normalise=False)
+        image_mfs = convolve2gaussres(model_mfs[None], xx, yy,
+                                    GaussPar[0], opts.nvthreads,
+                                    norm_kernel=False)[0]  # peak of kernel set to unity
+        image_mfs += residual_mfs
+        # convert pixel units to deg
+        GaussPar[0][0] *= cell_deg
+        GaussPar[0][1] *= cell_deg
+        hdr_mfs = add_beampars(hdr_mfs, GaussPar)
 
-    GaussPars = list(GaussPars)
-    for i, gp in enumerate(GaussPars):
-        GaussPars[i] = (gp[0]*cell_deg, gp[1]*cell_deg, gp[2])
-    GaussPars = tuple(GaussPars)
+        if any([i.isupper() for i in opts.outputs]):
+            GaussPars = fitcleanbeam(psf, level=0.5, pixsize=1.0)  # pixel units
+            cpsf = np.zeros(residual.shape, dtype=output_type)
+            for v in range(opts.nband):
+                cpsf[v] = Gaussian2D(xx, yy, GaussPars[v], normalise=False)
 
-    # init fits headers
-    radec = (dds[0].ra, dds[0].dec)
-    hdr_mfs = set_wcs(cell_deg, cell_deg, nx, ny, radec, np.mean(freq))
-    hdr = set_wcs(cell_deg, cell_deg, nx, ny, radec, freq)
+            image = np.zeros_like(model)
+            for b in range(nband):
+                image[b:b+1] = convolve2gaussres(model[b:b+1], xx, yy,
+                                                GaussPars[b], opts.nvthreads,
+                                                norm_kernel=False)  # peak of kernel set to unity
+                image[b] += residual[b]
+
+            for i, gp in enumerate(GaussPars):
+                GaussPars[i] = [gp[0]*cell_deg, gp[1]*cell_deg, gp[2]]
+
+            hdr = add_beampars(hdr, GaussPar, GaussPars)
+    else:
+        print('Warning, no psf in dds. '
+              'Unable to add resolution info or make restored image. ',
+              file=log)
 
     if 'm' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.model_mfs.fits', model_mfs, hdr_mfs,
+        save_fits(model_mfs,
+                  f'{basename}_{opts.postfix}.model_mfs.fits',
+                  hdr_mfs,
                   overwrite=opts.overwrite)
 
     if 'M' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.model.fits', model, hdr,
+        save_fits(model,
+                  f'{basename}_{opts.postfix}.model.fits',
+                  hdr,
                   overwrite=opts.overwrite)
 
-    # model does not get resolution info
-    hdr_mfs = add_beampars(hdr_mfs, GaussPar)
-    hdr = add_beampars(hdr, GaussPar, GaussPars)
-
     if 'r' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.residual_mfs.fits', residual_mfs, hdr_mfs,
+        save_fits(residual_mfs,
+                  f'{basename}_{opts.postfix}.residual_mfs.fits',
+                  hdr_mfs,
                   overwrite=opts.overwrite)
 
     if 'R' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.residual.fits', residual, hdr,
+        save_fits(residual,
+                  f'{basename}_{opts.postfix}.residual.fits',
+                  hdr,
                   overwrite=opts.overwrite)
 
-    if 'i' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.image_mfs.fits', image_mfs, hdr_mfs,
+    if 'd' in opts.outputs:
+        dirty_mfs = np.sum(dirty, axis=0)
+        save_fits(dirty_mfs,
+                  f'{basename}_{opts.postfix}.dirty_mfs.fits',
+                  hdr_mfs,
                   overwrite=opts.overwrite)
 
-    if 'I' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.image.fits', image, hdr,
+    if 'D' in opts.outputs:
+        dirty[fmask] /= wsums[fmask, None, None]/wsum
+        save_fits(dirty,
+                  f'{basename}_{opts.postfix}.dirty.fits',
+                  hdr,
                   overwrite=opts.overwrite)
 
-    if 'c' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.cpsf_mfs.fits', cpsf_mfs, hdr_mfs,
+    if 'i' in opts.outputs and psf is not None:
+        save_fits(image_mfs,
+                  f'{basename}_{opts.postfix}.image_mfs.fits',
+                  hdr_mfs,
                   overwrite=opts.overwrite)
 
-    if 'C' in opts.outputs:
-        save_fits(f'{basename}_{opts.postfix}.cpsf.fits', cpsf, hdr,
+    if 'I' in opts.outputs and psf is not None:
+        save_fits(image,
+                  f'{basename}_{opts.postfix}.image.fits',
+                  hdr,
+                  overwrite=opts.overwrite)
+
+    if 'c' in opts.outputs and psf is not None:
+        save_fits(cpsf_mfs,
+                  f'{basename}_{opts.postfix}.cpsf_mfs.fits',
+                  hdr_mfs,
+                  overwrite=opts.overwrite)
+
+    if 'C' in opts.outputs and psf is not None:
+        save_fits(cpsf,
+                  f'{basename}_{opts.postfix}.cpsf.fits',
+                  hdr,
                   overwrite=opts.overwrite)
 
     if opts.scheduler=='distributed':
